Refactor intersphinx code and add tests
- Remove sphobjinv._intersphinx
- Refactor still-needed functions to sphobjinv.cli.suggest
- Add tests for the CLI mapping-inference output
