+++ tests/test_cli_nonlocal.py
@@ -166,7 +166,7 @@ class TestSuggest:
 
     @pytest.mark.timeout(CLI_TEST_TIMEOUT * 4)
     def test_cli_suggest_from_url(self, misc_info, run_cmdline_test):
-        """Confirm name-only suggest works from URL."""
+        """Confirm reST-only suggest output works from URL."""
         with stdio_mgr() as (in_, out_, err_):
             run_cmdline_test(
                 [
@@ -180,29 +180,38 @@ class TestSuggest:
             )
             assert p_instance_of.search(out_.getvalue())
 
+    @pytest.mark.parametrize(
+        "url",
+        [
+            "http://sphobjinv.readthedocs.io/en/v2.0/modules/",
+            "http://sphobjinv.readthedocs.io/en/v2.0/modules/cmdline.html",
+            (
+                "http://sphobjinv.readthedocs.io/en/v2.0/modules/"
+                "cmdline.html#sphobjinv.cmdline.do_convert"
+            ),
+        ],
+    )
     @pytest.mark.timeout(CLI_TEST_TIMEOUT * 4)
-    def test_cli_suggest_from_dir_noanchor(self, run_cmdline_test):
-        """Confirm name-only suggest works from docpage URL."""
-        url = "http://sphobjinv.readthedocs.io/en/v2.0/modules/"
+    def test_cli_suggest_from_docset_urls(self, url, run_cmdline_test, check):
+        """Confirm reST-only suggest output works from URLs within a docset."""
         with stdio_mgr() as (in_, out_, err_):
             run_cmdline_test(["suggest", "-u", url, "inventory", "-at", "50"])
-            assert p_inventory.search(out_.getvalue())
+
+            check.is_true(p_inventory.search(out_.getvalue()))
+            check.is_in("LIKELY", err_.getvalue())
+            check.is_in(
+                "(http://sphobjinv.readthedocs.io/en/v2.0/, None)", err_.getvalue()
+            )
 
     @pytest.mark.timeout(CLI_TEST_TIMEOUT * 4)
-    def test_cli_suggest_from_page_noanchor(self, run_cmdline_test):
-        """Confirm name-only suggest works from docpage URL."""
-        url = "http://sphobjinv.readthedocs.io/en/v2.0/modules/cmdline.html"
+    def test_cli_suggest_from_objinv_url(self, run_cmdline_test, check):
+        """Confirm reST-only suggest works for direct objects.inv URL."""
+        url = "http://sphobjinv.readthedocs.io/en/v2.0/objects.inv"
         with stdio_mgr() as (in_, out_, err_):
             run_cmdline_test(["suggest", "-u", url, "inventory", "-at", "50"])
-            assert p_inventory.search(out_.getvalue())
 
-    @pytest.mark.timeout(CLI_TEST_TIMEOUT * 4)
-    def test_cli_suggest_from_page_withanchor(self, run_cmdline_test):
-        """Confirm name-only suggest works from docpage URL."""
-        url = (
-            "http://sphobjinv.readthedocs.io/en/v2.0/modules/"
-            "cmdline.html#sphobjinv.cmdline.do_convert"
-        )
-        with stdio_mgr() as (in_, out_, err_):
-            run_cmdline_test(["suggest", "-u", url, "inventory", "-at", "50"])
-            assert p_inventory.search(out_.getvalue())
+            check.is_true(p_inventory.search(out_.getvalue()))
+            check.is_in("PROBABLY", err_.getvalue())
+            check.is_in(
+                "(http://sphobjinv.readthedocs.io/en/v2.0/, None)", err_.getvalue()
+            )
