+++ tests/test_intersphinx.py
@@ -27,8 +27,7 @@ Sphinx |objects.inv| files.
 
 import pytest
 
-import sphobjinv as soi
-import sphobjinv._intersphinx as soi_isphx
+import sphobjinv.cli.suggest as soi_cli_suggest
 
 
 pytestmark = [pytest.mark.intersphinx, pytest.mark.local]
@@ -58,7 +57,9 @@ pytestmark = [pytest.mark.intersphinx, pytest.mark.local]
 )
 def test_strip_netloc_path(uri, trimmed, with_scheme):
     """Confirm that object URI trimming is working."""
-    assert trimmed == soi_isphx._strip_url_to_netloc_path(uri, with_scheme=with_scheme)
+    assert trimmed == soi_cli_suggest._strip_url_to_netloc_path(
+        uri, with_scheme=with_scheme
+    )
 
 
 @pytest.mark.parametrize(
@@ -72,66 +73,4 @@ def test_strip_netloc_path(uri, trimmed, with_scheme):
 )
 def test_extract_objinv_url_base(url, trimmed):
     """Confirm that inventory URL trimming is working."""
-    assert trimmed == soi_isphx.extract_objectsinv_url_base(url)
-
-
-@pytest.mark.parametrize(
-    ("web_url", "inv_url", "project", "mapping"),
-    [
-        (
-            "https://flask.palletsprojects.com/en/1.1.x/api/#flask.Config",
-            "https://flask.palletsprojects.com/en/1.1.x/objects.inv",
-            "flask",
-            ("https://flask.palletsprojects.com/en/1.1.x/", None),
-        ),
-        (
-            "https://docs.djangoproject.com/en/4.0/topics/cache/#memcached",
-            "https://docs.djangoproject.com/en/4.0/_objects/",
-            "django",
-            (
-                "https://docs.djangoproject.com/en/4.0/",
-                "https://docs.djangoproject.com/en/4.0/_objects/",
-            ),
-        ),
-        (
-            (
-                "https://docs.scipy.org/doc/numpy-1.13.0/reference/"
-                "arrays.interface.html#python-side"
-            ),
-            "https://docs.scipy.org/doc/numpy-1.13.0/objects.inv",
-            "numpy",
-            ("https://docs.scipy.org/doc/numpy-1.13.0/", None),
-        ),
-    ],
-    ids=(lambda arg: arg if (isinstance(arg, str) and "/" not in arg) else ""),
-)
-def test_infer_mapping(web_url, inv_url, project, mapping, res_path):
-    """Confirm intersphinx mapping inference works for select test cases.
-
-    These test(s) should continue to pass even if the various documentation sets
-    on the web are taken down. ``web_url`` and ``inv_url`` are chosen to be
-    valid and consistent with the versions of the |objects.inv| files stored
-    in the tests resource path `res_path`.
-
-    """
-    inv_path = res_path / f"objects_{project}.inv"
-    assert mapping == soi_isphx.infer_mapping(web_url, inv_url, soi.Inventory(inv_path))
-
-
-@pytest.mark.parametrize(
-    ("web_url", "project"),
-    [
-        (
-            "https://docs.djangoproject.com/en/4.0/topicXYZs/cache/#memcached",
-            "django",
-        )
-    ],
-    ids=(lambda arg: arg if (isinstance(arg, str) and "/" not in arg) else ""),
-)
-def test_no_matching_object(web_url, project, res_path):
-    """Confirm that no matching Inventory object is found when there shouldn't be."""
-    inv_path = res_path / f"objects_{project}.inv"
-    with pytest.raises(soi.SOIIsphxNoMatchingObjectError):
-        soi_isphx._extract_base_from_weburl_and_inventory(
-            web_url, soi.Inventory(inv_path)
-        )
+    assert trimmed == soi_cli_suggest.extract_objectsinv_url_base(url)
